+++ fat.c
@@ -26,7 +26,7 @@ void fatInit(){
    fatBootsector_t* bootaddr = (fatBootsector_t*) pointer;
    fatEbr32_t* ebraddr = (fatEbr32_t*) (pointer + sizeof(fatBootsector_t));
 
-   ide_read_sectors(0, 1, 0x000000800, 0x10, (uint32_t)bootaddr); //read VBR
+   ide_read_sectors(0, 1, 0x000000800, 0x10, (uint64_t)bootaddr); //read VBR
 
    bootsect = *bootaddr;
    ebrsect = *ebraddr;
@@ -40,11 +40,11 @@ void fatInit(){
    fat = (uint32_t*) kmalloc(ebrsect.fatSize * bootsect.bytesPerSect);
    //read the fat 
    mapPage((uint8_t*)0x1ff0000, (uint8_t*)0x1ff0000, 0x0);
-   ide_read_sectors(0, ebrsect.fatSize, 0x800 + bootsect.numReservedSects, 0x10, (uint32_t)fat); 
+   ide_read_sectors(0, ebrsect.fatSize, 0x800 + bootsect.numReservedSects, 0x10, (uint64_t)fat); 
 
 
    //load root dir
-   uint8_t status = ide_read_sectors(0, bootsect.sectsPerCluster, clusterToLba(ebrsect.rootCluster), 0x10, (uint32_t)rootaddr);
+   uint8_t status = ide_read_sectors(0, bootsect.sectsPerCluster, clusterToLba(ebrsect.rootCluster), 0x10, (uint64_t)rootaddr);
    rootFiles = rootaddr;
    kprintf("loading root dir at %h\n", (uint64_t)rootFiles);
    
@@ -122,14 +122,14 @@ int8_t openFile(char* fileName, fileHeader_t* loadAddr){
 //function takes in index of first cluster, then loads that cluster to loadAddr, increment with clustersize and load next cluster in chain
 uint8_t loadClusterChain(uint16_t firstCluster, fileHeader_t* loadAddr){
 
-   uint32_t addr = (uint32_t)loadAddr;
+   uint64_t addr = (uint64_t)loadAddr;
    uint32_t currentCluster = (uint32_t)firstCluster;
 
    uint8_t done = 0;
    do {
    
       //1. load cluster
-      ide_read_sectors(0, bootsect.sectsPerCluster, clusterToLba(currentCluster), 0x10, (uint32_t)addr);
+      ide_read_sectors(0, bootsect.sectsPerCluster, clusterToLba(currentCluster), 0x10, (uint64_t)addr);
 
       //2. increment loadAddr
       //addr += 0x200;
